refactor: ChatImgUrlDto 하나로 통합
