--- src/main/java/com/palettee/chat/controller/dto/request/ChatImgUrlRequest.java
@@ -1,15 +1,0 @@
-package com.palettee.chat.controller.dto.request;
-
-import com.palettee.chat.domain.Chat;
-import com.palettee.chat.domain.ChatImage;
-
-public record ChatImgUrlRequest(
-        String imgUrl
-) {
-    public ChatImage toEntityChatImage(Chat chat) {
-        return ChatImage.builder()
-                .chat(chat)
-                .imageUrl(this.imgUrl)
-                .build();
-    }
-}
